% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensure_exists.R
\name{ensure.exists}
\alias{ensure.exists}
\title{Download a file if it does not exist}
\usage{
ensure.exists(path, url)
}
\arguments{
\item{path}{The destination path on disk}

\item{url}{The source URL}
}
\description{
Download a file if it does not exist
}
