% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_lsm.R
\name{filepath2sample}
\alias{filepath2sample}
\title{Convert a vector of file paths into sample names}
\usage{
filepath2sample(paths, exts = c(".gz", ".fastq"))
}
\arguments{
\item{paths}{The file paths}

\item{exts}{The list of extensions to strip}
}
\description{
Convert a vector of file paths into sample names
}
