% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_cmdscale.R
\name{tidy_cmdscale}
\alias{tidy_cmdscale}
\title{Perform Classic Multi-dimensional scaling on a distance matrix, taking care of boilerplate code.}
\usage{
tidy_cmdscale(dist, k = 9, metadata = NULL, by = NULL)
}
\arguments{
\item{dist}{A distance matrix, either as a matrix or as a dist object}

\item{k}{The number of PCoA axes}

\item{metadata}{If not null, a dataframe (or tibble) describing additional sample metadata}

\item{by}{A vector informing `left_join` how to join (see the dplyr documentation)}
}
\description{
Returns a list containing the dataframe, preformatted axis labels, and eigenvalue magnitudes.
}
