% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_lsm.R
\name{read_lsm}
\alias{read_lsm}
\title{Reads a labelled square matrix to a correctly named matrix object}
\usage{
read_lsm(path, sep = "\\t", exts.to.remove = c(".gz", ".fastq", ".ct",
  ".msh"))
}
\arguments{
\item{path}{Path to LSM file on disk}

\item{sep}{The column separation character}
}
\description{
Reads a labelled square matrix to a correctly named matrix object
}
